pyproject.toml
[project]
name = "customer-support-agent"
version = "0.1.0"
description = ""
authors = [
    {name = "Caephas",email = "arinzeobidiegwu@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13,<4.0"
dependencies = [
    "fastapi[all] (>=0.115.8,<0.116.0)",
    "uvicorn (>=0.34.0,<0.35.0)",
    "firebase-admin (>=6.6.0,<7.0.0)",
    "langchain (>=0.3.17,<0.4.0)",
    "langgraph (>=0.2.69,<0.3.0)",
    "pydantic (>=2.10.6,<3.0.0)",
    "python-dotenv (>=1.0.1,<2.0.0)",
    "httpx (>=0.28.1,<0.29.0)",
    "langchain-community (>=0.3.16,<0.4.0)",
    "langchain-ollama (>=0.2.3,<0.3.0)",
    "pyjwt (>=2.10.1,<3.0.0)",
    "passlib (>=1.7.4,<2.0.0)",
    "simple-salesforce (>=1.12.6,<2.0.0)",
    "requests (>=2.32.3,<3.0.0)",
    "twilio (>=9.4.4,<10.0.0)",
    "slack-sdk (>=3.34.0,<4.0.0)",
    "sendgrid (>=6.11.0,<7.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
